it is hard to reliably remove the curly brackets in jsx props... so drop that.
while I managed to make it work in most common cases there are still some edge cases that can rise up...
for people that want that they can check out this branch.

But I do not want to deal with those problems... so it is better to not ship such functionality.